move new utils back to url-loader
