+++ packages/util/src/index.ts
@@ -9,12 +9,4 @@ export {
   type PollForProcessingImageOptions,
 } from "./lib/cloudinary.js";
 export { convertColorHexToRgb, testColorIsHex } from "./lib/colors.js";
-export {
-  encodeBase64,
-  entriesOf,
-  isArray,
-  objectHasKey,
-  sortByKey,
-  throwError,
-  type StringifiablePrimative,
-} from "./lib/util.js";
+export { encodeBase64, objectHasKey, sortByKey } from "./lib/util.js";
